+++ pyads/ads.py
@@ -394,13 +394,13 @@ def delete_route(adr):
     return adsDelRoute(adr.netIdStruct())
 
 
-def add_device_notification(adr, data_name, attr, callback, user_handle=None):
-    # type: (AmsAddr, str, NotificationAttrib, Callable, int) -> Optional[Tuple[int, int]]  # noqa: E501
+def add_device_notification(adr, data, attr, callback, user_handle=None):
+    # type: (AmsAddr, Union[str, Tuple[int, int], NotificationAttrib, Callable, int) -> Optional[Tuple[int, int]]  # noqa: E501
     """Add a device notification.
 
     :param pyads.structs.AmsAddr adr: AMS Address associated with the routing
         entry which is to be removed from the router.
-    :param str data_name: PLC storage address
+    :param Union[str, Tuple[int, int] data: PLC storage address as string or Tuple with index group and offset
     :param pyads.structs.NotificationAttrib attr: object that contains
         all the attributes for the definition of a notification
     :param callback: callback function that gets executed on in the event
@@ -416,7 +416,7 @@ def add_device_notification(adr, data_name, attr, callback, user_handle=None):
     """
     if port is not None:
         return adsSyncAddDeviceNotificationReqEx(
-            port, adr, data_name, attr, callback, user_handle
+            port, adr, data, attr, callback, user_handle
         )
 
     return None
@@ -898,11 +898,11 @@ class Connection(object):
                 self._port, self._adr, data_name, value, plc_datatype, handle=handle
             )
 
-    def add_device_notification(self, data_name, attr, callback, user_handle=None):
-        # type: (str, NotificationAttrib, Callable, int) -> Optional[Tuple[int, int]]
+    def add_device_notification(self, data, attr, callback, user_handle=None):
+        # type: (Union[str, Tuple[int, int]], NotificationAttrib, Callable, int) -> Optional[Tuple[int, int]]
         """Add a device notification.
 
-        :param str data_name: PLC storage address
+        :param Union[str, Tuple[int, int] data: PLC storage address as string or Tuple with index group and offset
         :param pyads.structs.NotificationAttrib attr: object that contains
             all the attributes for the definition of a notification
         :param callback: callback function that gets executed on in the event
@@ -945,7 +945,7 @@ class Connection(object):
         """
         if self._port is not None:
             notification_handle, user_handle = adsSyncAddDeviceNotificationReqEx(
-                self._port, self._adr, data_name, attr, callback, user_handle
+                self._port, self._adr, data, attr, callback, user_handle
             )
             return notification_handle, user_handle